Get paths reorganized
